+++ LearnLynks/src/main/java/com/project/LearnLynks/Services/CurriculumServiceImpl.java
@@ -114,7 +114,7 @@ public class CurriculumServiceImpl implements CurriculumService {
     }
 
     @Override
-    public ArchiveCurriculumResponse archive(int curriculumId) {
+    public ArchiveCurriculumResponse archive(Long curriculumId) {
 
         Curriculum curriculum = curriculumRepository.findById(curriculumId)
                 .orElseThrow(() -> new RuntimeException("Curriculum not found"));
