+++ LearnLynks/src/main/java/com/project/LearnLynks/dtos/request/DeleteCurriculumRequest.java
@@ -1,13 +1,13 @@
 package com.project.LearnLynks.dtos.request;
 
 public class DeleteCurriculumRequest {
-    public int getCurriculumId() {
+    public Long getCurriculumId() {
         return curriculumId;
     }
 
-    public void setCurriculumId(int curriculumId) {
+    public void setCurriculumId(Long curriculumId) {
         this.curriculumId = curriculumId;
     }
 
-    private int curriculumId;
+    private Long curriculumId;
 }
