+++ LearnLynks/src/main/java/com/project/LearnLynks/dtos/response/CreateCurriculumResponse.java
@@ -3,15 +3,16 @@ package com.project.LearnLynks.dtos.response;
 public class CreateCurriculumResponse {
     private String message;
 
-    public int getCurriculumId() {
+
+    public Long getCurriculumId() {
         return curriculumId;
     }
 
-    public void setCurriculumId(int curriculumId) {
+    public void setCurriculumId(Long curriculumId) {
         this.curriculumId = curriculumId;
     }
 
-    private int curriculumId;
+    private Long curriculumId;
 
     public String getMessage() {
         return message;
